rename path history to note
