+++ note/src/main/java/com/mediscreen/note/NoteApplication.java
@@ -1,4 +1,4 @@
-package com.mediscreen.history;
+package com.mediscreen.note;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
