+++ note/src/test/java/com/mediscreen/note/NoteApplicationTests.java
@@ -1,4 +1,4 @@
-package com.mediscreen.history;
+package com.mediscreen.note;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
